--- src/views/TrackerView.vue
@@ -1,14 +1,0 @@
-<template>
-    <AppWindow :isActive=true :top=0 :left=0 :width=800 :height=600 :minWidth=500 :minHeight=500
-        :isResizing="['r', 'rb', 'b', 'lb', 'l', 'lt', 't', 'rt']" :isButtonMaximized=true :isButtonMinimized=true
-        :isMaximized=false :maxWidth=1280 :maxHeight=720 windowId=1 title="Tracker" titleIcon="" @clickWindow=""
-        @clickDestroy="" @clickMin="">
-        <div class="flex justify-center">
-            test
-        </div>
-    </AppWindow>
-</template>
-
-<script setup lang="ts">
-import AppWindow from '@/components/AppWindow.vue';
-</script>
